+++ NEWS.md
@@ -8,6 +8,7 @@ If you are viewing this file on CRAN, please check [the latest news on GitHub](h
 - Fix: [#689](https://github.com/r-lidar/lidR/issues/689) attributes are not renamed when filtering a point cloud.
 - New: [#680](https://github.com/r-lidar/lidR/issues/680) it is now possible to use `plot(las, mapview = TRUE)`.
 - Fix: [#702](https://github.com/r-lidar/lidR/issues/702) `plot(las, breaks = <vector>)` now works with a vector of custom break points.
+- Fix: [#701](https://github.com/r-lidar/lidR/issues/701) `crown_metrics` now always remove invalid polygons if any.
 
 ## lidR v4.0.3 (Release date: 2023-03-09)